ai_interviewer/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.1.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-decouple==3.8
psycopg2-binary==2.9.9
alembic==1.13.2
langchain>=0.1.0
langchain-google-genai==2.1.6
langchain-huggingface==0.1.0
transformers==4.44.2
google-cloud-speech==2.27.0
pytest==8.3.3
celery==5.4.0
redis==5.1.1
google-auth>=2.0.0
langgraph>=0.2.0
python-multipart
